feat: return investment results through enrollment flow

+++ init-db/01_init.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS users (
     password    VARCHAR(255)    NOT NULL,
     name        VARCHAR(100)    NOT NULL,
     role        VARCHAR(20)     NOT NULL COMMENT 'STUDENT | INSTRUCTOR',
-    money       INT             NOT NULL DEFAULT 0,
+    money       DECIMAL(15,2)   NOT NULL DEFAULT 0,
     created_at  DATETIME(6),
     updated_at  DATETIME(6),
     PRIMARY KEY (id)
